+++ src/app/page.tsx
@@ -3,7 +3,7 @@
 import MainLayout from "@/layouts/MainLayout";
 import Notice from "@/components/main/Notice/Notice";
 import Schedule from "@/components/main/Schedule/Schedule";
-import Footer from "@/components/common/Footer/footer";
+import Footer from "@/components/common/Footer";
 import Video from "@/components/main/Video/Video";
 import Question from "@/components/main/Question/Question";
 import initMockAPI from "@/mocks";
