+++ src/app/signup/page.tsx
@@ -3,10 +3,10 @@
 import styled from "styled-components";
 import ButtonInput from "@/components/common/Input/ButtonInput";
 import PreviewInput from "@/components/common/Input/PreviewInput";
-import Button from "@/components/common/Button/button";
+import Button from "@/components/common/Button";
 import Image from "next/image";
 import Terms from "@/components/signup/Terms";
-import Column from "@/components/common/Flex/column";
+import { Column } from "@/components/common/Flex";
 import BaseLayout from "@/layouts/BaseLayout";
 import { color } from "@/styles/color";
 import { font } from "@/styles/font";
