--- src/components/common/Flex/type.ts
@@ -1,14 +1,0 @@
-import { ReactNode } from "react";
-
-export interface FlexPropsType {
-  children: ReactNode;
-  gap: string;
-  justifyContent?:
-    | "none"
-    | "center"
-    | "flex-end"
-    | "flex-start"
-    | "space-between";
-  alignItems?: "none" | "center" | "flex-end" | "flex-start" | "space-between";
-  width?: string;
-}
