+++ src/components/common/Header/index.tsx
@@ -3,8 +3,8 @@ import { ACCESS_KEY } from "@/constants/token";
 import { color } from "@/styles/color";
 import { useRouter } from "next/navigation";
 import styled from "styled-components";
-import Button from "../Button/button";
-import Row from "../Flex/row";
+import Button from "../Button";
+import { Row } from "../Flex";
 import LogoIcon from "../Icon/Logo";
 import NavigationBar from "./navigation";
 import Profile from "./profile";
