+++ src/components/common/Input/ButtonInput.tsx
@@ -1,4 +1,4 @@
-import Input from "./input";
+import Input from ".";
 import { color } from "@/styles/color";
 import { font } from "@/styles/font";
 import { InputPropsType } from "./type";
