+++ src/components/common/Input/PreviewInput.tsx
@@ -5,7 +5,7 @@ import { useState } from "react";
 import styled from "styled-components";
 import VisbilityOn from "../Icon/VisibilityOn";
 import VisbilityOff from "../Icon/VisibilityOff";
-import Message from "./message";
+import Message from "./Message";
 
 const PreviewInput = ({
   width = "320px",
