+++ src/components/main/Notice/Notice.tsx
@@ -4,7 +4,7 @@ import RightArrowIcon from "@/components/common/Icon/RightArrow";
 import NoticeItem from "./NoticeItem";
 import styled from "styled-components";
 import { useRouter } from "next/navigation";
-import Link from "@/components/common/Link/link";
+import Link from "@/components/common/Link";
 import { useMainNoticeListQuery } from "@/server/main/queries";
 
 const Notice = () => {
