+++ src/layouts/AppLayout.tsx
@@ -1,4 +1,4 @@
-import Header from "@/components/common/Header/header";
+import Header from "@/components/common/Header";
 import { color } from "@/styles/color";
 import { ReactNode } from "react";
 import styled from "styled-components";
